+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # UNRELEASED
+- [Kubernetes: 1.14.2](https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG-1.14.md#changelog-since-v1141)
 - [buildah: 1.8.2](https://github.com/containers/buildah/releases/tag/v1.8.2)
 - [podman: 1.3.1](https://github.com/containers/libpod/releases/tag/v1.3.1)
 - [skopeo: 0.1.36](https://github.com/containers/skopeo/releases/tag/v0.1.36)